locations/requirements.txt
flask-restful
flask-sqlalchemy
psycopg2-binary
